Moves work from signup to auth
